appoint dtype as int32 to fix a bug

+++ server/SegmentModel.py
@@ -29,7 +29,7 @@ class SegmentModel:
         
         self_trains_params = None
         self.inp = tf.placeholder(tf.int32, shape = [None, self.max_sentence_len], name = "input_placeholder")
-        self.loss(tf.zeros(shape=[1,80]), tf.zeros(shape=[1,80]))
+        self.loss(tf.zeros(shape=[1,80], dtype = int32), tf.zeros(shape=[1,80], dtype = int32))
         
     def length(self, data):
